+++ docs/src/examples/components/Text/States/TextExampleSuccess.tsx
@@ -1,4 +1,4 @@
-import React from 'react'
+import * as React from 'react'
 import { Text } from '@stardust-ui/react'
 
 const TextExampleSuccess = () => <Text success>Your action has completed successfully.</Text>