+++ test/specs/behaviors/listItemBehavior-test.tsx
@@ -1,15 +1,15 @@
 import { listItemBehavior } from 'src/lib/accessibility'
 
 describe('ListItemBehavior.ts', () => {
-  test('use SelectableListItemBehavior if selection prop is defined', () => {
+  test('use SelectableListItemBehavior if selectable prop is defined', () => {
     const property = {
-      selection: true,
+      selectable: true,
     }
     const expectedResult = listItemBehavior(property)
     expect(expectedResult.attributes.root.role).toEqual('option')
   })
 
-  test('use BasicListBehavior if selection prop is NOT defined', () => {
+  test('use BasicListBehavior if selectable prop is NOT defined', () => {
     const property = {}
     const expectedResult = listItemBehavior(property)
     expect(expectedResult.attributes.root.role).toEqual('listitem')